test: add tests for withCallback
